+++ external/anomalib/.gitignore
