+++ external/anomalib/anomaly_classification/src/callbacks/inference.py
