+++ external/anomalib/anomaly_classification/src/callbacks/model_monitor.py
