+++ external/anomalib/anomaly_classification/src/callbacks/progress.py
