+++ external/anomalib/anomaly_classification/src/configuration.yaml
