+++ external/anomalib/anomaly_classification/src/openvino.py
