+++ external/anomalib/anomaly_classification/src/utils/__init__.py
