+++ external/anomalib/anomaly_classification/src/utils/data.py
