+++ external/anomalib/tests/helpers/__init__.py
