+++ external/anomalib/tests/helpers/config.py
