+++ external/anomalib/tests/helpers/dataset.py
